+++ ChangeLog.md
@@ -3,7 +3,7 @@
 ## 0.7.0.0 -- 2016-11-09
 
 * Split openssl support to [mysql-haskell-openssl](hackage.haskell.org/package/mysql-haskell-openssl).
-* Expose `Database.MySQL.Connection` module due to this split, it's shouldn't be used by user directly.
+* Expose `Database.MySQL.Connection` module due to this split, it shouldn't be used by user directly.
 
 ## 0.6.0.0 -- 2016-10-25